Add files via upload
This needs additional features which I have not added as we discussed prior. We need to add a linked list to categorize and output final search result. I will go through the implementation in tomorrows meeting. the search method can be improved but only after the implementation of the sort function which you are working on. See you tomorrow :)